User commit

+++ src/main/java/com/example/demo/service/Impl/StudentServiceImpl.java
@@ -21,6 +21,7 @@ public class StudentServiceImpl implements StudentService{
     }
     @Override
     public String DeleteData(int id){
+        if(i)
         student.deleteById(id);
         return "Deleted Succesfully";
     }
